src/backend/requirements.txt
# ── backend/requirements.txt ──────────────────────────────────────────────────
fastapi==0.111.0
uvicorn[standard]==0.29.0
psycopg2-binary==2.9.9
numpy==1.26.4
pydantic==2.7.0
python-dotenv==1.0.1
requests==2.31.0

